+++ src/app/[language]/brokers/alpaca/backtest/page.tsx
@@ -1,0 +1,60 @@
+import * as React from 'react';
+import "@/app/globals.css";
+import { getServerSession } from "next-auth/next"
+import { authOptions } from "@/app/lib/auth";
+import { getDictionary } from '@/app/lib/dictionaries/dictionaries';
+import { LanguageProps } from '@/models/common/language-props';
+import { AlpacaUserSettingsService } from '@/service/alpaca/user-settings-service';
+import TestSettingsForm from '@/components/alpaca/test-settings-form';
+
+
+export default async function AlpacaSettingsPage({ params }: LanguageProps) {
+
+  const dict = await getDictionary(params.language)
+  const session = await getServerSession(authOptions)
+
+
+
+
+  if (session !== null) {
+    if (session.user) {
+      console.log("session", session)
+
+    }
+  }
+
+  if (session === null) {
+
+    return (
+      <div className="content-container flex flex-col items-center w-full pt-[10%]">
+        <p className='text-slate-800 font-semibold text-2xl mt-6'>Sie müssen angemeldet sein um diese Funktion zu nutzen.</p>
+      </div>
+    )
+
+
+  }
+
+
+  // eingabe test settings
+  // übersicht tests
+  // Anzeige test ergebnisse
+  // Anzeige positions
+  // Anzeige  chart mit positionen/trades
+
+  return (
+
+    <div className="flex flex-row gap-4 w-full">
+      <div className="flex-1 content-container">
+        <TestSettingsForm />
+      </div>
+      <div className="flex-1 content-container">
+        test
+      </div>
+      <div className="flex-1 content-container">
+        test
+      </div>
+
+    </div>
+
+  );
+}