+++ src/models/alpaca/alpaca-user-settings.ts
@@ -1,0 +1,6 @@
+export interface AlpacaUserSettings {
+    email: string;
+    symbols: string;
+    alpacaKey: string;
+    alpacaSecret: string;
+}