+++ src/types.ts
@@ -13,7 +13,7 @@ declare module 'next-auth/jwt' {
 }
 
 declare module 'next-auth' {
-    interface User {
+    interface User  {
       id?: string;
       role?: string;
       name: string;