+++ src/utilities.ts
@@ -1,0 +1,6 @@
+export function firstOrDefault<T>(array: T[], defaultValue: T): T {
+    if (!array) {
+        return defaultValue;
+    }
+    return array.length > 0 ? array[0] : defaultValue;
+}